// src/api/models/pet.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Category } from '../models/category';
import { Tag } from '../models/tag';
export interface Pet {
  category?: Category;
  id?: number;
  name: string;
  photoUrls: Array<string>;

  /**
   * pet status in the store
   */
  status?: 'available' | 'pending' | 'sold';
  tags?: Array<Tag>;
}